fix(controller): 将评论和收藏计数返回类型从Integer改为Long
修改NoteCommentController和NoteFavoriteController中计数接口的返回类型，以支持更大的数值范围

+++ backend/src/main/java/com/wyl/note/controller/NoteCommentController.java
@@ -49,8 +49,8 @@ public class NoteCommentController {
 
     @Operation(summary = "获取评论数")
     @GetMapping("/count")
-    public Result<Integer> getCommentCount(@RequestParam Long noteId) {
-        Integer count = noteCommentService.getCommentCount(noteId);
+    public Result<Long> getCommentCount(@RequestParam Long noteId) {
+        Long count = noteCommentService.getCommentCount(noteId);
         return Result.success(count);
     }
 }
